requirements.txt
streamlit
speechrecognition
google-cloud-texttospeech
google-cloud-speech
openai
pinecone-client
requests
python-dotenv